requirements.txt
apt-xapian-index==0.47
beautifulsoup4==4.4.1
blinker==1.3
chardet==2.3.0
click==6.7
cloud-init==0.7.8
command-not-found==0.3
configobj==5.0.6
cryptography==1.2.3
dateutils==0.6.6
html5lib==0.999
idna==2.0
Jinja2==2.8
jsonpatch==1.10
jsonpointer==1.9
language-selector==0.1
lxml==3.5.0
MarkupSafe==0.23
Mastodon.py==1.0.6
nltk==3.0.5
numpy==1.9.0
oauthlib==2.0.1
prettytable==0.7.2
prosaic==6.1.1
psycopg2==2.6.1
pyasn1==0.1.9
pycurl==7.43.0
pygobject==3.20.0
pyhocon==0.3.29
PyJWT==1.3.0
pyparsing==2.1.10
pyserial==3.0.1
pystache==0.5.4
python-apt==1.1.0b1
python-dateutil==2.6.0
python-debian==0.1.27
python-systemd==231
pytz==2017.2
PyYAML==3.11
requests==2.13.0
requests-oauthlib==0.8.0
six==1.10.0
SQLAlchemy==1.0.12
ssh-import-id==5.5
svgwrite==1.1.10
tildetown==1.0.1
tweepy==3.5.0
ufw==0.35
unattended-upgrades==0.1
urllib3==1.13.1
virtualenv==15.0.1
